% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{confint.rcr}
\alias{confint.rcr}
\title{Confidence intervals for RCR results}
\usage{
\method{confint}{rcr}(object, parm, level = 0.95, citype = "conservative", ...)
}
\arguments{
\item{object}{An object of class "\code{rcr}", usually the result of a call
to \code{\link[=rcr]{rcr()}}.}

\item{parm}{An optional specification of which parameters are to be given
confidence intervals, either a vector of numbers or a vector
of names. Available parameters include "\code{rcInf}", "\code{effectInf}",
"\code{rc0}", "\code{effectL}", "\code{effectH}", and "\code{effect}".
If missing, all parameters are considered.}

\item{level}{The confidence level required, on a scale
of 0 to 1. Default is 0.95.}

\item{citype}{The confidence interval type: "\code{conservative}" (the default)
"\code{upper}", "\code{lower}" or "\code{Imbens-Manski}".}

\item{...}{Additional arguments (not used).}
}
\value{
\code{confint.rcr()} returns a matrix with rows corresponding to
the elements of \code{parm} and columns giving upper and lower
confidence limits for each parameter
}
\description{
\code{confint.rcr()} computes confidence intervals for one or more parameters
in an RCR model.
}
\examples{
# Estimate the model
result <- rcr(weight ~ Time | Diet, ChickWeight)
# Use confint() to produce the confidence intervals
confint(result)
}
\seealso{
\code{\link[=rcr]{rcr()}} to estimate the model, \code{\link[=effect_test]{effect_test()}} to perform
hypothesis tests on the causal effect.
}
