% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{effect_test}
\alias{effect_test}
\title{Hypothesis testing for RCR causal effect}
\usage{
effect_test(object, h0 = 0)
}
\arguments{
\item{object}{an object of class "rcr", usually the result of a call
to \code{rcr}.}

\item{h0}{the value of the parameter under the null hypothesis.  Default
is zero.}
}
\value{
\code{effect_test()} returns the p-value associated with the
specified null hypothesis.
}
\description{
\code{effect_test()} allows a user to test a point null hypothesis on the
(interval identified) causal effect parameter in the RCR model.
}
\details{
The hypothesis test is based on inverting the Imbens-Manski confidence
interval rather than on a specific test statistic or set of critical
values.  For example, we reject the null hypothesis at 5\% significance
if the value under the null is outside of the 95\% Imbens-Manski
confidence interval.
}
\seealso{
\code{\link[=rcr]{rcr()}}, \code{\link[=confint.rcr]{confint.rcr()}}
}
