% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{install_rcrpy}
\alias{install_rcrpy}
\title{Install rcrbounds Python package}
\usage{
install_rcrpy(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}
}
\description{
The rcrbounds R package is a wrapper to a Python
package, also called "rcrbounds". This function
installs the required Python package from the
Python Package Index (PyPI), and may need to be
run once before using the functions in the R
package.
}
\details{
For most users, this function can be run once with the default
arguments and then ignored.
}
\seealso{
\code{\link[=rcr]{rcr()}}
\code{\link[reticulate:py_install]{reticulate::py_install()}} which this function wraps.
}
