% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{plot.rcr}
\alias{plot.rcr}
\title{Plot method for RCR results}
\usage{
\method{plot}{rcr}(
  x,
  xlab = expression(beta[x]),
  ylab = expression(lambda),
  xlim = c(-50, 50),
  ylim = NULL,
  col = "black",
  lty = c("solid", "dotted", "dashed"),
  ...
)
}
\arguments{
\item{x}{An object of class "\code{rcr}", usually the result of a call
to \code{rcr}}

\item{xlab, ylab}{Axis labels for x-axis and y-axis, respectively}

\item{xlim, ylim}{Range of values for x-axis and y-axis, respectively.}

\item{col}{A vector of colors to be used for the model elements.  \code{col[1]}
will be the color of the \code{lambda(beta_x)} function, \code{col[2]} will
be the color of the vertical line depicting \code{effectInf}, and
\code{col[3]}will be the color of the horizontal line depicting \code{rcInf}.
If fewer than 3 colors are provided, \code{plot.rcr} will cycle
through the provided colors. Colors can be provided as simple
names, or in any other form described in the "Color Specification"
section of the documentation for \code{\link[=par]{par()}}.}

\item{lty}{A vector of line types to be used for the model elements. \code{lty[1]}
will be the type of the \code{lambda(beta_x)} function, \code{lty[2]} will
be the type of the vertical line depicting \code{effectInf}, and
\code{lty[3]}will be the type of the horizontal line depicting \code{rcInf}.
If fewer than 3 line types are provided, \code{plot.rcr} will cycle
through the provided line types. Line types can be provided as simple
names, or in any other form described in the "Line Type Specification"
section of the documentation for \code{\link[=par]{par()}}.}

\item{...}{Additional graphical parameters to be passed on to \code{\link[=plot.default]{plot.default()}}.}
}
\value{
\code{plot.rcr()} returns its input object \code{x} invisibly.
}
\description{
\code{plot.rcr()} plots the main results of estimating an RCR model. The plot
will display three elements: the \code{lambda(beta_x)} function, the value of \code{effectInf},
and the value of \code{rcInf}.
}
\details{
It is recommended that you create an initial plot with the default
values of \code{xlim} and \code{ylim}, and then adjust these values as needed
to zero in on the portion of the graph that is of interest to you.

It is also recommended that you use \code{plot.rcr()} during the same R
session in which you created the \code{rcr} object using the \code{\link[=rcr]{rcr()}}
command, and ideally shortly after you create the object. The
reason for this is that it uses Python code embedded in the
\code{rcr} object, and this code will typically not be available if the
current Python session ends (e.g., by an end in the R session).
}
\seealso{
\code{\link[=rcr]{rcr()}} to estimate the model
}
