% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{print.rcr}
\alias{print.rcr}
\title{Print method for RCR results}
\usage{
\method{print}{rcr}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{rcr}", usually the result of a call
to \code{rcr}}

\item{...}{Additional arguments to be passed on to \code{print()},
for example \verb{digits=}.}
}
\value{
\code{print.rcr()} returns its input object \code{x} invisibly.
}
\description{
\code{print.rcr()} prints the main results of estimating an RCR model
}
\seealso{
\code{\link[=rcr]{rcr()}}, \code{\link[=print.summary.rcr]{print.summary.rcr()}}
}
