% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{summary.rcr}
\alias{summary.rcr}
\alias{print.summary.rcr}
\title{Summary method for RCR results}
\usage{
\method{summary}{rcr}(object, level = 0.95, citype = "conservative", ...)

\method{print}{summary.rcr}(x, ...)
}
\arguments{
\item{object}{an object of class "rcr", usually the result of a call
to \code{rcr}.}

\item{level}{the confidence level required.}

\item{citype}{the confidence interval type: "\code{conservative}" (the default)
"\code{upper}", "\code{lower}" or "\code{Imbens-Manski}".}

\item{...}{Additional arguments (e.g., \code{digits}) to be
passed to \code{print()}}

\item{x}{an object of class "summary.rcr", usually the result
of a call to \code{summary.rcr}.}
}
\value{
\code{summary.rcr()} returns an rcr.summary object, which
is a list with the following components:
}
\description{
\code{summary.rcr()} produces a tabular summary of rcr model
results estimated by the \code{\link[=rcr]{rcr()}} function.
}
\seealso{
\code{\link[=rcr]{rcr()}}.
}
