% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{vcov.rcr}
\alias{vcov.rcr}
\title{Calculate covariance matrix for RCR coefficients}
\usage{
\method{vcov}{rcr}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{rcr}", usually the result of a call
to \code{\link[=rcr]{rcr()}}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{vcov()} returns the variance-covariance matrix of the main
point-identified parameters of the rcr model object.
}
\description{
\code{vcov()} returns the variance-covariance matrix of the main parameters
of an rcr model object.
}
\examples{
# Estimate the model
result <- rcr(weight ~ Time | Diet, ChickWeight)
# Use coef() to recover the coefficients
coef(result)
# Use vcov() to recover the covariance matrix
vcov(result)
}
\seealso{
\code{\link[=rcr]{rcr()}} to estimate the model, \code{\link[=coef]{coef()}} to retrieve coefficient
estimates
}
