% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{confint.rcr}
\alias{confint.rcr}
\title{Confidence intervals for RCR results}
\usage{
\method{confint}{rcr}(object, parm, level = 0.95, citype = "conservative", ...)
}
\arguments{
\item{object}{an object of class "rcr", usually the result of a call
to \code{rcr}.}

\item{parm}{a specification of which parameters are to be given
confidence intervals, either a vector of numbers or a vector
of names. Available parameters include "\code{rcInf}", "\code{effectInf}",
"\code{rc0}", "\code{effectL}", "\code{effectH}", and "\code{effect}".
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{citype}{the confidence interval type: "\code{conservative}" (the default)
"\code{upper}", "\code{lower}" or "\code{Imbens-Manski}".}

\item{...}{additional arguments (not used).}
}
\value{
\code{confint.rcr()} returns a matrix with columns giving
upper and lower confidence limits for each parameter
}
\description{
Computes confidence intervals for one or more parameters in an RCR model.
}
\seealso{
\code{\link[=rcr]{rcr()}}
}
