% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{install_rcrpy}
\alias{install_rcrpy}
\title{Install rcrbounds Python package}
\usage{
install_rcrpy(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}
}
\value{
\code{install_rcrpy()} returns whatever is returned by the
call to \link[reticulate:py_install]{reticulate::py_install}, usually \code{NULL}.
}
\description{
\code{install_rcrpy()} is a utility function to install the rcrbounds
Python package from the Python Package Index (PyPI). This function
may need to be run once before using the rcrbounds R package.
For most users, this function can be run once with the default
arguments and then ignored.
}
\seealso{
\code{\link[=rcr]{rcr()}} to estimate an RCR model,
\code{\link[reticulate:py_install]{reticulate::py_install()}} which this function wraps.
}
