% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{rcr}
\alias{rcr}
\title{Estimate an RCR model}
\usage{
rcr(
  formula,
  data,
  subset,
  weights,
  na.action,
  model = TRUE,
  pyobj = TRUE,
  cluster = NULL,
  rc_range = c(0, 1),
  vceadj = 1
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that
can be coerced to that class): a symbolic description
of the model to be fitted. The details of model specification
are given under 'Details'.}

\item{data}{an optional data frame, list or environment (or object
coercible by as.data.frame to a data frame) containing the
variables in the model. If not found in data, the variables
are taken from environment(formula), typically the environment
from which rcr is called.}

\item{subset}{an optional vector specifying a subset of observations
to be used in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be NULL or a numeric vector.}

\item{na.action}{a function which indicates what should happen when
the data contain NAs. The default is set by the na.action setting
of options, and is na.fail if that is unset. The ‘factory-fresh’
default is na.omit. Another possible value is NULL, no action.
Value na.exclude can be useful.}

\item{model, pyobj}{logicals. If TRUE the corresponding intermediate components
of the fit (the model frame, the Python object) are returned.}

\item{cluster}{An optional vector defining groups for cluster-robust
covariance matrix estimation. Should be NULL or a numeric
vector. See also ‘Details’,}

\item{rc_range}{An optional numeric vector of length two indicating
the assumed range for the relative correlation parameter
lambda.\code{-Inf} is allowed for the lower bound, and \code{Inf} is
allowed for the upper bound.}

\item{vceadj}{An optional adjustment factor to perform degrees-of-freedom
adjustments for the estimated covariance matrix.  That is,
the estimated covariance matrix will be multiplied by \code{vceadj}.}
}
\value{
\code{rcr()} returns an object of class "rcr", which is
a list containing the following components:
\describe{
\item{\code{coefficients}}{a named vector of coefficients}
\item{\code{cov.unscaled}}{the estimated covariance matrix for \code{coefficients}}
\item{\code{na.action}}{(where relevant) information returned by model.frame on the special handling of NAs..}
\item{\code{xlevels}}{(where relevant) a record of the levels of the factors used in fitting..}
\item{\code{call}}{the matched call.}
\item{\code{terms}}{the terms object used.}
\item{\code{model}}{if requested (the default), the model frame used.}
\item{\code{weights}}{(where relevant) the specified weights.}
\item{\code{cluster}}{(where relevant) the specified cluster identifier.}
\item{\code{pyobj}}{a pointer to the Python object returned
by \code{\link[=rcr.fit]{rcr.fit()}}. }
}
}
\description{
\code{rcr()} estimates the linear casual effect of a scalar explanatory
variable on a scalar outcome, under a relative correlation restriction
of the form described in Krauth (2016).
}
\details{
Most of the calculation is done in an external Python function which
is called by \code{rcr()}. You may need to use the \code{\link[=install_rcrpy]{install_rcrpy()}} function
to install the package the first time you wish to use this function.
}
\examples{
f1 <- 1
}
\seealso{
\code{\link[=rcr.fit]{rcr.fit()}},
}
