% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\name{vcov.rcr}
\alias{vcov.rcr}
\title{Calculate covariance matrix for RCR coefficients}
\usage{
\method{vcov}{rcr}(object, ...)
}
\arguments{
\item{object}{an object of class "rcr", usually the result of a call
to \code{rcr}.}

\item{...}{Additional arguments (not used).}
}
\value{
A symmetric 5-by-5 matrix with row and column names corresponding
to the parameter names in \code{coef(object)}.
}
\description{
\code{vcov()} Returns the variance-covariance matrix of the main parameters
of the rcr model object.
}
